+++ python/GafferImageUI/ResizeUI.py
@@ -115,7 +115,7 @@ Gaffer.Metadata.registerNode(
 
 			"description",
 			"""
-			The pixel used when transforming the image. Each
+			The filter used when transforming the image. Each
 			filter provides different tradeoffs between sharpness and
 			the danger of aliasing or ringing.
 			""",
@@ -123,6 +123,7 @@ Gaffer.Metadata.registerNode(
 			"plugValueWidget:type", "GafferUI.PresetsPlugValueWidget",
 
 			"preset:Default", "",
+			"preset:Nearest", "nearest",
 
 		] + list( itertools.chain(
 
@@ -131,6 +132,22 @@ Gaffer.Metadata.registerNode(
 			*[ ( "preset:" + x.title(), x ) for x in GafferImage.FilterAlgo.filterNames() if x != "disk" ]
 
 		) ),
+
+		"filterDeep" : [
+
+			"description",
+			"""
+			By default, we use a nearest filter for deep images, because
+			accurately filtering deeps can be extremely expensive, and
+			produces images that are too large to store on disk. If you
+			know what you're doing, turn this on to do accurate filtering
+			( the only manageable use case is probably if you immediately
+			do a merge or holdout after resampling, so the a minimum
+			amount of processing needs to happen to the huge data of the
+			resampled deep ).
+			""",
+
+		],
 
 	}